clean out and make create profile layout function based
