+++ services/frontend-v3/src/components/layouts/appLayout/AppLayout.jsx
@@ -6,7 +6,6 @@ function AppLayout(props) {
     <AppLayoutView
       changeLanguage={props.changeLanguage}
       keycloak={props.keycloak}
-      history={props.history}
       displaySideBar={props.displaySideBar}
       sideBarContent={props.sideBarContent}
     >
