+++ services/frontend-v3/src/pages/ProfileCreate.jsx
@@ -16,7 +16,6 @@ class ProfileCreate extends React.Component {
         changeLanguage={this.props.changeLanguage}
         keycloak={this.props.keycloak}
         history={this.props.history}
-        displaySideBar={true}
         step={this.props.match.params.step}
       ></CreateProfileLayout>
     );
